% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{add_variable_process}
\alias{add_variable_process}
\title{add_variable_process}
\usage{
add_variable_process(add)
}
\arguments{
\item{add}{A data.frame contained address variables.}
}
\description{
This function is not intended to be used by end user.
}
