% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_derivation.R
\name{address_varieble}
\alias{address_varieble}
\title{address_varieble}
\usage{
address_varieble(df, address_cols = NULL, address_pattern = NULL,
  parallel = TRUE)
}
\arguments{
\item{df}{A data.frame.}

\item{address_cols}{Variables of address,}

\item{address_pattern}{Regular expressions, used to match address variable names.}

\item{parallel}{Logical, parallel computing. Default is TRUE.}
}
\description{
This function is not intended to be used by end user.
}
