% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_missing.R
\name{analysis_nas}
\alias{analysis_nas}
\title{Missing Analysis}
\usage{
analysis_nas(dat, class_var = FALSE, nas_rate = NULL, na_vars = NULL,
  mat_nas_shadow = NULL, dt_nas_random = NULL, ...)
}
\arguments{
\item{dat}{A data.frame with independent variables and target variable.}

\item{class_var}{Logical, nas analysis of the nominal variables. Default is TRUE.}

\item{nas_rate}{A list contains nas rate of each variable.}

\item{na_vars}{Names of variables which contain nas.}

\item{mat_nas_shadow}{A shadow matrix of variables which contain nas.}

\item{dt_nas_random}{A data.frame with random nas imputation.}

\item{...}{Other parameters.}
}
\value{
A data.frame with outliers analysis for each variable.
}
\description{
#' \code{analysis_nas} is the function for missing value analysis
}
