% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_outliers.R
\name{analysis_outliers}
\alias{analysis_outliers}
\title{Outliers Analysis}
\usage{
analysis_outliers(dat, target, x, lof = NULL)
}
\arguments{
\item{dat}{A data.frame with independent variables and target variable.}

\item{target}{The name of target variable.}

\item{x}{The name of variable to process.}

\item{lof}{Outliers of each variable detected by \code{outliers_detection}.}
}
\value{
A data.frame with outliers analysis for each variable.
}
\description{
#' \code{analysis_outliers} is the function for outliers analysis.
}
