% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/essential_algorithms.R
\name{auc_value}
\alias{auc_value}
\title{auc_value
\code{auc_value} is for get best lambda required in lasso_filter. This function required in \code{lasso_filter}}
\usage{
auc_value(target, prob)
}
\arguments{
\item{target}{Vector of target.}

\item{prob}{A list of redict probability or score.}
}
\value{
Lanmbda value
}
\description{
auc_value
\code{auc_value} is for get best lambda required in lasso_filter. This function required in \code{lasso_filter}
}
