% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_derivation.R
\name{city_varieble}
\alias{city_varieble}
\title{city_varieble}
\usage{
city_varieble(df = df, city_cols = NULL, city_pattern = "city$",
  city_class = city_class, parallel = TRUE)
}
\arguments{
\item{df}{A data.frame.}

\item{city_cols}{Variables of city,}

\item{city_pattern}{Regular expressions, used to match city variable names. Default is "city$".}

\item{city_class}{Class or levels of cities.}

\item{parallel}{Logical, parallel computing. Default is TRUE.}
}
\description{
This function is used for city variables derivation.
}
