% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_derivation.R
\name{city_varieble_process}
\alias{city_varieble_process}
\title{Processing of Address Variables}
\usage{
city_varieble_process(df_city, x, city_class)
}
\arguments{
\item{df_city}{A data.frame.}

\item{x}{Variables of city,}

\item{city_class}{Class or levels of cities.}
}
\description{
This function is not intended to be used by end user.
}
