% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualization.R
\name{cohort_table_plot}
\alias{cohort_table_plot}
\alias{cohort_plot}
\title{cohort_table_plot
\code{cohort_table_plot} is for ploting cohort(vintage) analysis table.}
\usage{
cohort_table_plot(cohort_dat)

cohort_plot(cohort_dat)
}
\arguments{
\item{cohort_dat}{A data.frame generated by \code{cohort_analysis}.}
}
\description{
This function is not intended to be used by end user.
}
