% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_breaks.R
\name{cut_equal}
\alias{cut_equal}
\title{Generating Initial Equal Size Sample Bins}
\usage{
cut_equal(dat_x, g = 10, sp_values = list(-1, "Unknown"))
}
\arguments{
\item{dat_x}{A vector of an variable x.}

\item{g}{numeric, number of initial bins for equal_bins.}

\item{sp_values}{a list of special value. Default: list(-1, "Unknown")}
}
\description{
\code{cut_equal} is used to generate initial breaks for equal frequency binning.
}
\examples{
\dontrun{
#equal sample size breaks
equ_breaks = cut_equal(dat = UCICreditCard[, "PAY_AMT2"], g = 10)

#tree breaks
tree_control = list(p = 0.02, cp = 0.000001, xval = 5, maxdepth = 10)
tree_breaks = get_tree_breaks(dat = UCICreditCard, x = "MARRIAGE", 
target = "default.payment.next.month", tree_control = tree_control)

# select best bins
bins_control = list(bins_num = 10, bins_pct = 0.02, b_chi = 0.02, b_odds = 0.1, b_psi = 0.05,
b_gb = 15, mono = 0.2, gb_psi = 0.1, kc = 5)

select_best_class(dat = UCICreditCard, x = "MARRIAGE", breaks = tree_breaks,
target = "default.payment.next.month", occur_time = "apply_date", 
sp_values = NULL, bins_control = bins_control)

select_best_breaks(dat = UCICreditCard, x = "PAY_AMT2",
breaks = equ_breaks, target = "default.payment.next.month",
occur_time = "apply_date", 
sp_values = NULL, bins_control = bins_control)
}
}
\seealso{
\code{\link{get_breaks}}, \code{\link{get_breaks_all}},\code{\link{get_tree_breaks}}
}
