% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process_tools.R
\name{cv_split}
\alias{cv_split}
\title{Stratified Folds}
\usage{
cv_split(dat, k = 5, occur_time = NULL, seed = 46)
}
\arguments{
\item{dat}{A data.frame.}

\item{k}{k is an integer specifying the number of folds.}

\item{occur_time}{time variable for creating OOT folds. Default is NULL.}

\item{seed}{A seed. Default is 46.}
}
\value{
a list of indices
}
\description{
this function creates stratified folds for cross validation.
}
\examples{
sub = cv_split(UCICreditCard, k = 30)[[1]]
dat = UCICreditCard[sub,]
}
