% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_proc_tools.R
\name{date_cut}
\alias{date_cut}
\title{date cut point}
\usage{
date_cut(dat_time, pct = 0.7)
}
\arguments{
\item{dat_time}{time vectors.}

\item{pct}{the percent of cutting. Default: 0.7.}
}
\value{
A Date.
}
\description{
\code{date_cut} is  a small function to get date point.
}
\examples{
\dontrun{
date_cut(dat_time = lendingclub$issue_d, pct = 0.8)
}
}
