% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{de_percent}
\alias{de_percent}
\title{Recovery Percent Format}
\usage{
de_percent(x, digits = 2)
}
\arguments{
\item{x}{Character with percent formant.}

\item{digits}{Number of digits.Default: 2.}
}
\value{
x without percent format.
}
\description{
\code{de_percent} is  a small function for recoverying percent format..
}
\examples{
de_percent("24\%")
}
