% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_derivation_tools.R
\name{derived_interval}
\alias{derived_interval}
\title{derived_interval}
\usage{
derived_interval(dat_s, interval_type = c("cnt_interval",
  "time_interval"))
}
\arguments{
\item{dat_s}{A data.frame contained only predict variables.}

\item{interval_type}{Available of c("cnt_interval", "time_interval")}
}
\description{
This function is not intended to be used by end user.
}
