% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_derivation_tools.R
\name{derived_partial_acf}
\alias{derived_partial_acf}
\title{derived_partial_acf}
\usage{
derived_partial_acf(dat_s)
}
\arguments{
\item{dat_s}{A data.frame}
}
\description{
This function is not intended to be used by end user.
}
