% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_derivation_tools.R
\name{derived_pct}
\alias{derived_pct}
\title{derived_pct}
\usage{
derived_pct(dat_s, pct_type = "total_pct")
}
\arguments{
\item{dat_s}{A data.frame contained only predict variables.}

\item{pct_type}{Available of "total_pct"}
}
\description{
This function is not intended to be used by end user.
}
