% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_proc_tools.R
\name{digits_num}
\alias{digits_num}
\title{digits_num}
\usage{
digits_num(num)
}
\arguments{
\item{num}{A sample of numeric variable.}
}
\value{
A number of digits
}
\description{
\code{digits_num} is for caculating optimal digits number for numeric variables.
}
\examples{
\dontrun{
digits_num(3.1415296)
}
}
