% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{entry_rate_max}
\alias{entry_rate_max}
\title{Max Percent of Unique Values}
\usage{
entry_rate_max(dat, rt = 0.95, note = TRUE)
}
\arguments{
\item{dat}{A data frame with x and target.}

\item{rt}{The maximum percent of unique value.}

\item{note}{Logical.Outputs info.Default is TRUE.}
}
\value{
A data.frame
}
\description{
\code{entry_rate_max} is the function to encode variables which unique value  rate is  more than 95% as 1,0
}
\examples{
datss = entry_rate_max(dat = UCICreditCard , rt  = 0.98)
}
