% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleansing.R
\name{entry_rate_na}
\alias{entry_rate_na}
\title{Max Percent of Missing Value}
\usage{
entry_rate_na(dat, nr = 0.95, note = FALSE)
}
\arguments{
\item{dat}{A data frame with x and target.}

\item{nr}{The maximum percent of NAs.}

\item{note}{Logical.Outputs info.Default is TRUE.}
}
\value{
A data.frame
}
\description{
\code{entry_rate_na} is the function to recode variables which NAs or  miss value rate is more than 95%.
}
\examples{
datss = entry_rate_na(dat = lendingclub[1:1000, ], nr = 0.98)
}
