% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_customer_segmentation.R
\name{euclid_dist}
\alias{euclid_dist}
\title{euclid_dist}
\usage{
euclid_dist(x, y, margin = 1)
}
\arguments{
\item{x}{A list}

\item{y}{A list}

\item{margin}{rows or cols}
}
\description{
This function is not intended to be used by end user.
}
