% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/essential_algorithms.R
\name{eval_auc}
\alias{eval_auc}
\alias{eval_ks}
\alias{eval_tnr}
\alias{eval_lift}
\title{Functions of xgboost feval}
\usage{
eval_auc(preds, dtrain)

eval_ks(preds, dtrain)

eval_tnr(preds, dtrain)

eval_lift(preds, dtrain)
}
\arguments{
\item{preds}{A list of predict probability or score.}

\item{dtrain}{Matrix of x predictors.}
}
\value{
List of best value
}
\description{
\code{eval_auc} ,\code{eval_ks} ,\code{eval_lift},\code{eval_tnr} is for getting best params of xgboost.
}
