% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.r
\docType{data}
\name{ewm_data}
\alias{ewm_data}
\title{Entropy Weight Method Data}
\format{
A data frame with 10 rows and 13 variables.
}
\description{
This data is for Entropy Weight Method examples.
}
\keyword{datasets}
