% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_training.R
\name{gbm_params}
\alias{gbm_params}
\title{GBM Parameters}
\usage{
gbm_params(n.trees = 1000, interaction.depth = 6, shrinkage = 0.01,
  bag.fraction = 0.5, train.fraction = 0.7, n.minobsinnode = 30,
  cv.folds = 5, ...)
}
\arguments{
\item{n.trees}{Integer specifying the total number of trees to fit. This is equivalent to the number of iterations and the number of basis functions in the additive expansion. Default is 100.}

\item{interaction.depth}{Integer specifying the maximum depth of each tree(i.e., the highest level of variable interactions allowed) . A value of 1 implies an additive model, a value of 2 implies a model with up to 2 - way interactions, etc. Default is 1.}

\item{shrinkage}{a shrinkage parameter applied to each tree in the expansion. Also known as the learning rate or step - size reduction; 0.001 to 0.1 usually work, but a smaller learning rate typically requires more trees. Default is 0.1 .}

\item{bag.fraction}{the fraction of the training set observations randomly selected to propose the next tree in the expansion. This introduces randomnesses into the model fit. If bag.fraction < 1 then running the same model twice will result in similar but different fits. gbm uses the R random number generator so set.seed can ensure that the model can be reconstructed. Preferably, the user can save the returned gbm.object using save. Default is 0.5 .}

\item{train.fraction}{The first train.fraction * nrows(data) observations are used to fit the gbm and the remainder are used for computing out-of-sample estimates of the loss function.}

\item{n.minobsinnode}{Integer specifying the minimum number of observations in the terminal nodes of the trees. Note that this is the actual number of observations, not the total weight.}

\item{cv.folds}{Number of cross - validation folds to perform. If cv.folds > 1 then gbm, in addition to the usual fit, will perform a cross - validation, calculate an estimate of generalization error returned in cv.error.}

\item{...}{Other parameters}
}
\value{
A list of parameters.
}
\description{
\code{gbm_params} is the list of parameters to train a GBM using in  \code{\link{training_model}}.
}
\details{
See details at: \code{gbm}
}
\seealso{
\code{\link{training_model}}, \code{\link{lr_params}}, \code{\link{xgb_params}}, \code{\link{rf_params}}
}
