% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_selection.R
\name{get_auc_ks_lambda}
\alias{get_auc_ks_lambda}
\title{get_auc_ks_lambda
\code{get_auc_ks_lambda} is for get best lambda required in lasso_filter. This function required in \code{lasso_filter}}
\usage{
get_auc_ks_lambda(lasso_model, x_test, y_test, save_data = FALSE,
  plot_show = TRUE, file_name = NULL, dir_path = tempdir())
}
\arguments{
\item{lasso_model}{A lasso model genereted by glmnet.}

\item{x_test}{A matrix of test dataset with x.}

\item{y_test}{A matrix of y test dataset with y.}

\item{save_data}{Logical, save results in locally specified folder. Default is FALSE}

\item{plot_show}{Logical, if TRUE plot the results. Default is TRUE.}

\item{file_name}{The name for periodically saved results files. Default is NULL.}

\item{dir_path}{The path for periodically saved results files.}
}
\value{
Lanmbda values with max K-S and AUC.
}
\description{
get_auc_ks_lambda
\code{get_auc_ks_lambda} is for get best lambda required in lasso_filter. This function required in \code{lasso_filter}
}
\seealso{
\code{\link{lasso_filter}}, \code{\link{get_sim_sign_lambda}}
}
