% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_filter_lasso.R
\name{get_best_lambda}
\alias{get_best_lambda}
\title{get_best_lambda
\code{plot_theme} is for get Best lambda required in lasso_filter. This function required in \code{lasso_filter}}
\usage{
get_best_lambda(lasso_model, lasso_model_cv, type.measure = "auc",
  sim_sign = "negtive")
}
\arguments{
\item{lasso_model}{A lasso model genereted by glmnet.}

\item{lasso_model_cv}{A cv lasso model genereted by cv.glmnet}

\item{type.measure}{Default is  "auc".}

\item{sim_sign}{Default is "negtive". This is related to pos_plag. If pos_flag equals 1 or bad, the value must be set to negetive. If pos_flag equals 0 or good, the value must be set to positive.}
}
\value{
Four lanmbda values with different thresholds.
}
\description{
get_best_lambda
\code{plot_theme} is for get Best lambda required in lasso_filter. This function required in \code{lasso_filter}
}
\details{
lambda.min give a model with the best performance but the least number of independent variable.
lambda.lse give the simplest model in the range of a standard deviation
lambda.05se give the simplest model in the range of a half standard deviation
lambda.sim_sign give the model with the same positive or negetive coefficients of all variables.
lambda.sim_sign give the model with the same positive or negetive coefficients of all variables.
}
