% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_score_card.R
\name{get_logistic_coef}
\alias{get_logistic_coef}
\title{get logistic coef}
\usage{
get_logistic_coef(lg_model, file_name = "LR_coef",
  dir_path = "./model", save_data = TRUE)
}
\arguments{
\item{lg_model}{An object of logistic model.}

\item{file_name}{The name for periodically saved coefficient file.  Default is "LR_coef".}

\item{dir_path}{The Path for periodically saved coefficient file. Default is "./model".}

\item{save_data}{Logical, save the result or not. Default is TRUE.}
}
\value{
A data.frame with logistic coefficients.
}
\description{
\code{get_logistic_coef} is  for geting logistic coefficient.
}
\examples{
\dontrun{
get_logistic_coef(lg_model = model, train_woe = train_woe, dat_test = dat_test)
}
}
