% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_missing.R
\name{get_median}
\alias{get_median}
\title{get central value.}
\usage{
get_median(x, weight_avg = NULL)
}
\arguments{
\item{x}{A vector or list.}

\item{weight_avg}{avarage weigth to calculate means.}
}
\description{
This function is not intended to be used by end user.
}
