% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_missing.R
\name{get_nas_random}
\alias{get_nas_random}
\title{get_nas_random}
\usage{
get_nas_random(dat)
}
\arguments{
\item{dat}{A data.frame contained only predict variables.}
}
\description{
This function is not intended to be used by end user.
}
