% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_score_card.R
\name{get_score_card}
\alias{get_score_card}
\title{Score Card}
\usage{
get_score_card(lg_model, target, bins_table, a = 600, b = 50,
  file_name = "LR_Score_Card", dir_path = "./model",
  save_data = TRUE)
}
\arguments{
\item{lg_model}{An object of glm model.}

\item{target}{The name of target variable.}

\item{bins_table}{a data.frame generated by \code{\link{get_bins_table}}}

\item{a}{Base line of score.}

\item{b}{Numeric.Increased scores from doubling Odds.}

\item{file_name}{The name for periodically saved scorecard file. Default is "LR_Score_Card".}

\item{dir_path}{The path for periodically saved scorecard file. Default is "./model"}

\item{save_data}{Logical, save results in locally specified folder. Default is TRUE}
}
\value{
scorecard.
}
\description{
\code{get_score_card} is  for generating a stardard scorecard
}
\examples{
\dontrun{
get_score_card(lg_model, bins_table, target = "target")
}
}
