% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_missing.R
\name{get_shadow_nas}
\alias{get_shadow_nas}
\title{get_shadow_nas}
\usage{
get_shadow_nas(dat)
}
\arguments{
\item{dat}{A data.frame contained only predict variables.}
}
\description{
This function is not intended to be used by end user.
}
