% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_selection.R
\name{get_sim_sign_lambda}
\alias{get_sim_sign_lambda}
\title{get_sim_sign_lambda
\code{get_sim_sign_lambda} is for get Best lambda required in lasso_filter. This function required in \code{lasso_filter}}
\usage{
get_sim_sign_lambda(lasso_model, sim_sign = "negtive")
}
\arguments{
\item{lasso_model}{A lasso model genereted by glmnet.}

\item{sim_sign}{Default is "negtive". This is related to pos_plag. If pos_flag equals 1 or 1, the value must be set to negetive. If pos_flag equals 0 or 0, the value must be set to positive.}
}
\value{
Lanmbda value
}
\description{
get_sim_sign_lambda
\code{get_sim_sign_lambda} is for get Best lambda required in lasso_filter. This function required in \code{lasso_filter}
}
\details{
lambda.sim_sign give the model with the same positive or negetive coefficients of all variables.
}
