% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process_tools.R
\name{\%alike\%}
\alias{\%alike\%}
\title{%alike%
Fuzzy String matching}
\usage{
x \%alike\% y
}
\arguments{
\item{x}{A string.}

\item{y}{A string.}
}
\value{
Logical.
}
\description{
%alike%
Fuzzy String matching
}
\examples{
"xyz"  \%alike\% "xy"
}
