% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process_tools.R
\name{\%islike\%}
\alias{\%islike\%}
\title{%islike%
Fuzzy String matching}
\usage{
x \%islike\% y
}
\arguments{
\item{x}{A string.}

\item{y}{A string.}
}
\value{
Logical.
}
\description{
%islike%
Fuzzy String matching
}
\examples{
 "xyz"  \%islike\% "yz$"
}
