% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_selection.R
\name{high_cor_selector}
\alias{high_cor_selector}
\title{Compare the two highly correlated variables}
\usage{
high_cor_selector(
  cor_mat,
  p = 0.95,
  x_list = NULL,
  com_list = NULL,
  retain = TRUE
)
}
\arguments{
\item{cor_mat}{A correlation matrix.}

\item{p}{The threshold of high correlation.}

\item{x_list}{Names of independent variables.}

\item{com_list}{A data.frame with important values of each variable. eg : IV_list.}

\item{retain}{Logical, output selected variables, if FALSE, output filtered variables.}
}
\value{
A list of selected variables.
}
\description{
\code{high_cor_selector} is function for comparing the two highly correlated variables, select a variable with the largest IV value.
}
