% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process_tools.R
\name{is_date}
\alias{is_date}
\title{is_date}
\usage{
is_date(x)
}
\arguments{
\item{x}{list or vectors}
}
\value{
A Date.
}
\description{
\code{is_date} is a small function for distinguishing time formats
}
\examples{
is_date(lendingclub$issue_d)
}
