% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_outliers.R
\name{k_distance}
\alias{k_distance}
\title{k_distance
\code{k_distance} is for returning an object in which each column
This function is not intended to be used by end user.}
\usage{
k_distance(dataset, neighbors)
}
\arguments{
\item{dataset}{A data.frame contained only predict variables.}

\item{neighbors}{A data.frame contained neighbors of each obs.}
}
\description{
k_distance
\code{k_distance} is for returning an object in which each column
This function is not intended to be used by end user.
}
