% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_outliers.R
\name{k_distance_neighbor}
\alias{k_distance_neighbor}
\title{k_distance_neighbor}
\usage{
k_distance_neighbor(x, dat, k)
}
\arguments{
\item{x}{The name of variable to process.}

\item{dat}{A data.frame contained only predict variables.}

\item{k}{Number of neighbors for LOF.}
}
\description{
\code{k_distance_neighbor}  is used for returning knn distance neighbors.
This function is not intended to be used by end user.
}
