% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_missing.R
\name{knn_nas_imp}
\alias{knn_nas_imp}
\title{Imputate nas using KNN}
\usage{
knn_nas_imp(dat, x, nas_rate = NULL, mat_nas_shadow = NULL,
  dt_nas_random = NULL, k = 10, scale = FALSE, method = "median")
}
\arguments{
\item{dat}{A data.frame with independent variables.}

\item{x}{The name of variable to process.}

\item{nas_rate}{A list contains nas rate of each variable.}

\item{mat_nas_shadow}{A shadow matrix of variables which contain nas.}

\item{dt_nas_random}{A data.frame with random nas imputation.}

\item{k}{Number of neighbors of each obs which x is missing.}

\item{scale}{Logical.Standardization of variable.}

\item{method}{The methods of imputation by knn."median" is knn imputation by k neighbors median.}
}
\description{
This function is not intended to be used by end user.
}
