% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/essential_algorithms.R
\name{ks_value}
\alias{ks_value}
\title{ks_value}
\usage{
ks_value(target, prob)
}
\arguments{
\item{target}{Vector of target.}

\item{prob}{A list of redict probability or score.}
}
\value{
KS value
}
\description{
\code{ks_value} is for get K-S value for a prob or score.
}
