% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/essential_algorithms.R
\name{lift_value}
\alias{lift_value}
\title{lift_value}
\usage{
lift_value(target, prob)
}
\arguments{
\item{target}{Vector of target.}

\item{prob}{A list of predict probability or score.}
}
\value{
Max lift value
}
\description{
\code{lift_value} is for getting max lift value for a prob or score.
}
