% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/essential_algorithms.R
\name{local_outlier_factor}
\alias{local_outlier_factor}
\title{local_outlier_factor
\code{local_outlier_factor}  is function for calculating the lof factor for a data set using knn
This function is not intended to be used by end user.}
\usage{
local_outlier_factor(dat, k = 10)
}
\arguments{
\item{dat}{A data.frame contained only predict variables.}

\item{k}{Number of neighbors for LOF.Default is 10.}
}
\description{
local_outlier_factor
\code{local_outlier_factor}  is function for calculating the lof factor for a data set using knn
This function is not intended to be used by end user.
}
