% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_outliers.R
\name{local_reachability_density}
\alias{local_reachability_density}
\title{local_reachability_density}
\usage{
local_reachability_density(distance_data, k)
}
\arguments{
\item{distance_data}{A matrix with distance of each observation.}

\item{k}{Number of neighbors for LOF.}
}
\description{
\code{local_reachability_density} is used for calculating local reachability density.
}
\details{
This function is not intended to be used by end user.
}
