% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{love_color}
\alias{love_color}
\title{love_color}
\usage{
love_color(color = "dark_cyan")
}
\arguments{
\item{color}{The name of colors.}
}
\description{
\code{love_color} is for get plots for a  variable.
}
\examples{
love_color("dark_cyan")
}
