% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process_tools.R
\name{min_max_norm}
\alias{min_max_norm}
\title{Min Max Normalization}
\usage{
min_max_norm(x)
}
\arguments{
\item{x}{Vector}
}
\value{
Normalized vector
}
\description{
\code{min_max_norm} is for normalizing each column vector of matrix 'x' using min_max normalization
}
\examples{
dat_s = apply(UCICreditCard[,12:14], 2, min_max_norm)
}
