% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{model_key_index}
\alias{model_key_index}
\title{model_key_index
\code{model_key_index} is for get plots for a  variable.}
\usage{
model_key_index(tb_pred)
}
\arguments{
\item{tb_pred}{A table generated by code{\link{ks_table}}}
}
\description{
model_key_index
\code{model_key_index} is for get plots for a  variable.
}
\examples{
model_key_index(tb_pred)
}
