% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process_tools.R
\name{multi_left_jion}
\alias{multi_left_jion}
\title{multi_left_jion}
\usage{
multi_left_jion(df_list, key_dt = NULL, by = NULL)
}
\arguments{
\item{df_list}{A list of datasets.}

\item{key_dt}{Name or index of Key table to left join.}

\item{by}{Name of Key columns to join.}
}
\description{
\code{multi_left_jion} is for left jion a list of datasets fast.
}
\examples{
save_dt(UCICreditCard,"UCICreditCard", tempdir())
}
