% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process_tools.R
\name{multi_left_join}
\alias{multi_left_join}
\title{multi_left_join}
\usage{
multi_left_join(..., df_list = list(...), key_dt = NULL, by = NULL)
}
\arguments{
\item{...}{Datasets need join}

\item{df_list}{A list of datasets.}

\item{key_dt}{Name or index of Key table to left join.}

\item{by}{Name of Key columns to join.}
}
\description{
\code{multi_left_join} is for left jion a list of datasets fast.
}
\examples{
multi_left_join(UCICreditCard[1:10, 1:10], UCICreditCard[1:10, c(1,8:14)],
UCICreditCard[1:10, c(1,20:25)], by = "ID")
}
