% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process_tools.R
\name{n_char}
\alias{n_char}
\title{The length of a string.}
\usage{
n_char(string)
}
\arguments{
\item{string}{A string.}
}
\value{
A numeric vector giving number of characters (code points) in each
   element of the character vector. Missing string have missing length.
}
\description{
Returns the number of "code points", in a string.
}
\examples{
n_char(letters)
n_char(NA)
}
