% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{null_blank_na}
\alias{null_blank_na}
\title{Encode NAs}
\usage{
null_blank_na(dat, miss_values = NULL, note = TRUE)
}
\arguments{
\item{dat}{A data frame with x and target.}

\item{miss_values}{Other extreme value might be used to represent missing values, e.g: -9999, -9998. These miss_values will be encoded to -1 or "Unknown".}

\item{note}{Logical.Outputs info.Default is TRUE.}
}
\value{
A data.frame
}
\description{
\code{null_blank_na} is the function to  replace null ,NULL, blank or other missing vaules with NA.
}
\examples{
datss = null_blank_na(dat = UCICreditCard, miss_values =list(-1,-2))
}
