% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_outliers.R
\name{outliers_detection}
\alias{outliers_detection}
\title{Outliers Detection
\code{outliers_detection} is for outliers detecting using Kmeans and Local Outlier Factor (lof)}
\usage{
outliers_detection(dat, x, kc = 3, kn = 5)
}
\arguments{
\item{dat}{A data.frame with independent variables.}

\item{x}{The name of variable to process.}

\item{kc}{Number of  clustering centers for Kmeans}

\item{kn}{Number of neighbors for LOF.}
}
\value{
Outliers of each variable.
}
\description{
Outliers Detection
\code{outliers_detection} is for outliers detecting using Kmeans and Local Outlier Factor (lof)
}
