% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/essential_algorithms.R
\name{p_ij}
\alias{p_ij}
\alias{e_ij}
\title{Entropy}
\usage{
p_ij(x)

e_ij(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector of entropy.
}
\description{
This function is not intended to be used by end user.
}
