% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_reduce_PCA.R
\name{PCA_reduce}
\alias{PCA_reduce}
\title{PCA Dimension Reduction}
\usage{
PCA_reduce(train = train, test = NULL, mc = 0.9)
}
\arguments{
\item{train}{A data.frame with independent variables and target variable.}

\item{test}{A data.frame of test data.}

\item{mc}{Threshold of cumulative imp.}
}
\description{
\code{PCA_reduce} is used for PCA reduction of high demension data .
}
\examples{
\dontrun{
num_x_list = get_names(dat = UCICreditCard, types = c('numeric'),  
ex_cols = "ID$|date$|default.payment.next.month$", get_ex = FALSE)
 PCA_dat = PCA_reduce(train = UCICreditCard[num_x_list])
}
}
