% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualization.R
\name{plot_bar}
\alias{plot_bar}
\title{Plot Bar}
\usage{
plot_bar(dat, x, breaks = NULL, g = 10, cut_bin = "equal_width")
}
\arguments{
\item{dat}{A data.frame with independent variables and target variable.}

\item{x}{The name of an independent variable.}

\item{breaks}{Breaks points of x.}

\item{g}{Number of initial breakpoints for equal frequency binning.}

\item{cut_bin}{'equal_width' or 'equal_depth'}
}
\description{
You can use the \code{plot_bar} to produce the barplot.
}
\examples{
plot_bar(dat = lendingclub, x = "grade")
}
