% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualization.R
\name{plot_box}
\alias{plot_box}
\title{Plot Box}
\usage{
plot_box(
  dat,
  y,
  x = NULL,
  g = 5,
  colors_x = c(love_color(type = "deep"), love_color(type = "light"), love_color(type =
    "shallow"), love_color(type = "dark"))
)
}
\arguments{
\item{dat}{A data.frame with independent variables and target variable.}

\item{y}{The name of target variable.}

\item{x}{The name of an independent variable.}

\item{g}{Number of initial breakpoints for equal frequency binning.}

\item{colors_x}{colors of x.}
}
\description{
You can use the \code{plot_box} to produce boxplot.
}
\examples{
plot_box(lendingclub, x = "grade", y = "installment", g = 7)
}
