% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualization.R
\name{plot_colors}
\alias{plot_colors}
\alias{color_ramp_palette}
\title{Plot Colors}
\usage{
plot_colors(colors)

color_ramp_palette(colors)
}
\arguments{
\item{colors}{A vector of colors.}
}
\description{
You can use the \code{plot_colors} to show colors on the graph device.
}
\examples{
plot_colors(rgb(158,122,122, maxColorValue = 255 ))
}
