% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{plot_theme}
\alias{plot_theme}
\title{plot_theme}
\usage{
plot_theme(legend.position = "top", angle = 30, legend_size = 7,
  axis_size_y = 8, axis_size_x = 8, axis_title_size = 10,
  title_size = 11, title_vjust = 0, title_hjust = 0,
  linetype = "dotted", face = "bold")
}
\arguments{
\item{legend.position}{see details at: code{legend.position}}

\item{angle}{see details at:  code{axis.text.x}}

\item{legend_size}{see details at:  code{legend.text}}

\item{axis_size_y}{see details at:  code{axis.text.y}}

\item{axis_size_x}{see details at:  code{axis.text.x}}

\item{axis_title_size}{see details at:  code{axis.title.x}}

\item{title_size}{see details at:  code{plot.title}}

\item{title_vjust}{see details at:  code{plot.title}}

\item{title_hjust}{see details at:  code{plot.title}}

\item{linetype}{see details at:  code{panel.grid.major}}

\item{face}{see details at:  code{axis.title.x}}
}
\description{
\code{plot_theme} is a simper wrapper of theme for ggplot2.
}
\details{
see details at: code{theme}
}
