% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_training.R
\name{pred_score}
\alias{pred_score}
\title{pred_score}
\usage{
pred_score(model, dat, x_list = NULL, bins_table = NULL,
  obs_id = NULL, sp_values = list(-1, "Missing"), save_data = TRUE,
  woe_name = TRUE)
}
\arguments{
\item{model}{Model generated by \code{\link{training_model}}.}

\item{dat}{Dataframe of new data.}

\item{x_list}{Into the model variables.}

\item{bins_table}{a data.frame generated by \code{\link{get_bins_table}}}

\item{obs_id}{The name of ID of observations or key variable of data. Default is NULL.}

\item{sp_values}{Special values.}

\item{save_data}{Logical, save the result or not. Default is FALSE.}

\item{woe_name}{Logical. Whether woe variable's name contains 'woe'.Default is TRUE.}
}
\value{
new scores.
}
\description{
\code{pred_score} is for using trained models to predict new data.
}
\seealso{
\code{\link{training_model}}, \code{\link{lr_params}}, \code{\link{xgb_params}}, \code{\link{rf_params}}
}
