% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_outliers.R
\name{process_outliers}
\alias{process_outliers}
\alias{outliers_kmeans_lof}
\title{Process Outliers}
\usage{
process_outliers(dat, target, ex_cols = NULL, kc = 3, kn = 5,
  x_list = NULL, parallel = FALSE, note = TRUE, process = TRUE,
  dir_path = "./variable")

outliers_kmeans_lof(dat, x, target = NULL, kc = 3, kn = 5,
  note = TRUE, process = TRUE, dir_path = "./variable")
}
\arguments{
\item{dat}{Dataset with independent variables and target variable.}

\item{target}{The name of target variable.}

\item{ex_cols}{A list of excluded variables. Regular expressions can also be used to match variable names. Default is NULL.}

\item{kc}{Number of  clustering centers for Kmeans}

\item{kn}{Number of neighbors for LOF.}

\item{x_list}{Names of independent variables.}

\item{parallel}{Logical, parallel computing.}

\item{note}{Logical, outputs info. Default is TRUE.}

\item{process}{Logical, process outliers, not just analysis.}

\item{dir_path}{The path for periodically saved outliers analysis file. Default is "./variable".}

\item{x}{The name of variable to process.}
}
\value{
A data frame with outliers process to all the variables.
}
\description{
\code{outliers_kmeans_lof} Outliers processing using Kmeans and Local Outlier Factor (lof)
\code{process_outliers} is a simpler wrapper for \code{outliers_kmeans_lof}.
}
\examples{
\dontrun{
# load germancredit dat
dat(UCI_Credit_Card)
dat_out = process_outliers(dat_clean, 
target = "default.payment.next.month",
ex_cols = "date$", kc = 3, kn = 10, parallel = FALSE)
}

}
