% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_proc_tools.R
\name{quick_as_df}
\alias{quick_as_df}
\title{quick_as_df}
\usage{
quick_as_df(df_list)
}
\arguments{
\item{df_list}{A list of data.}
}
\value{
packages installed and library,
}
\description{
\code{quick_as_df} is function for fast dat frame  transfromation.
}
\examples{
\dontrun{
UCICreditCard = quick_as_df(UCICreditCard)
}
}
