% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process_tools.R
\name{re_code}
\alias{re_code}
\title{re_code
\code{re_code} search for matches to argument pattern within each element of a character vector:}
\usage{
re_code(x, codes)
}
\arguments{
\item{x}{Variable to recode.}

\item{codes}{A data.frame of original value & recode value}
}
\description{
re_code
\code{re_code} search for matches to argument pattern within each element of a character vector:
}
\examples{
SEX  = sample(c("F","M"),1000,replace = TRUE)
codes= data.frame(ori_value = c('F','M'), code = c(0,1) )
SEX_re = re_code(SEX,codes)
}
