% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process_tools.R
\name{re_name}
\alias{re_name}
\title{Rename}
\usage{
re_name(dat, oldname = c(), newname = c())
}
\arguments{
\item{dat}{A data frame with vairables to rename.}

\item{oldname}{Old names of vairables.}

\item{newname}{New names of vairables.}
}
\value{
data with new variable names.
}
\description{
\code{re_name} is  for renaming variables.
}
\examples{
dt = re_name(dat = UCICreditCard, "default.payment.next.month" , "target")
names(dt['target'])
}
