% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process_tools.R
\name{read_data}
\alias{read_data}
\alias{check_data_format}
\title{Read data}
\usage{
read_data(
  path,
  pattern = NULL,
  encoding = "unknown",
  header = TRUE,
  sep = "auto",
  stringsAsFactors = FALSE,
  select = NULL,
  drop = NULL,
  nrows = Inf
)

check_data_format(path)
}
\arguments{
\item{path}{Path to file or file name in working directory & path to file.}

\item{pattern}{An optional regular expression. Only file names which match the regular expression will be returned.}

\item{encoding}{Default is "unknown". Other possible options are "UTF-8" and "Latin-1".}

\item{header}{Does the first data line contain column names?}

\item{sep}{The separator between columns.}

\item{stringsAsFactors}{Logical. Convert all character columns to factors?}

\item{select}{A vector of column names or numbers to keep, drop the rest.}

\item{drop}{A vector of column names or numbers to drop, keep the rest.}

\item{nrows}{The maximum number of rows to read.}
}
\description{
\code{read_data} is for loading data, formats like csv, txt,data and so on.
}
