% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process_tools.R
\name{read_dt}
\alias{read_dt}
\alias{check_data_format}
\title{Read data}
\usage{
read_dt(path, encoding = "unknown", header = "auto", sep = "auto",
  stringsAsFactors = FALSE)

check_data_format(path)
}
\arguments{
\item{path}{Path to file or file name in working directory & path to file.}

\item{encoding}{Default is "unknown". Other possible options are "UTF-8" and "Latin-1".}

\item{header}{Does the first data line contain column names?}

\item{sep}{The separator between columns.}

\item{stringsAsFactors}{Logical. Convert all character columns to factors?}
}
\description{
\code{read_dt} is for loading data.
}
