% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_filter_high_cor.R
\name{reduce_high_cor}
\alias{reduce_high_cor}
\title{Compare the two highly correlated variables}
\usage{
reduce_high_cor(cor_mat, p = 0.9, x_list = NULL, com_list = NULL,
  retain = TRUE)
}
\arguments{
\item{cor_mat}{A correlation matrix.}

\item{p}{The threshold of high correlation.}

\item{x_list}{Names of independent variables.}

\item{com_list}{A data.frame with important values of each variable. eg : IV_list.}

\item{retain}{Logical, output selected variables, if FALSE, output filtered variables.}
}
\value{
A list of selected variables.
}
\description{
\code{reduce_high_cor} is function for comparing the two highly correlated variables, select a variable with the largest IV value.
}
