% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleansing.R
\name{remove_duplicated}
\alias{remove_duplicated}
\title{Remove Duplicated Observations}
\usage{
remove_duplicated(dat = dat, obs_id = NULL, occur_time = NULL,
  target = NULL, note = FALSE)
}
\arguments{
\item{dat}{A data frame with x and target.}

\item{obs_id}{The name of ID of observations. Default is NULL.}

\item{occur_time}{The name of occur time of observations.Default is NULL.}

\item{target}{The name of target variable.}

\item{note}{Logical.Outputs info.Default is TRUE.}
}
\value{
A data.frame
}
\description{
\code{remove_duplicated} is the function to remove duplicated observations
}
\examples{
datss = remove_duplicated(dat = UCICreditCard,
target = "default.payment.next.month",
obs_id = "ID", occur_time =  "apply_date")
}
