% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{replace_value}
\alias{replace_value}
\alias{replace_value_x}
\title{Replace Value}
\usage{
replace_value(
  dat = dat,
  x_list = NULL,
  x_pattern = NULL,
  replace_dat,
  MARGIN = 2,
  VALUE = if (MARGIN == 2) colnames(replace_dat) else rownames(replace_dat),
  RE_NAME = TRUE,
  parallel = FALSE
)

replace_value_x(
  dat,
  x,
  replace_dat,
  MARGIN = 2,
  VALUE = if (MARGIN == 2) colnames(replace_dat) else rownames(replace_dat),
  RE_NAME = TRUE
)
}
\arguments{
\item{dat}{A data.frame.}

\item{x_list}{Names of variables to replace value.}

\item{x_pattern}{Regular expressions, used to match variable names.}

\item{replace_dat}{A data.frame contains value to replace.}

\item{MARGIN}{A vector giving the subscripts which the function will be applied over. E.g., for a matrix 1 indicates rows, 2 indicates columns, c(1, 2) indicates rows and columns. Where X has named dimnames, it can be a character vector selecting dimension names.}

\item{VALUE}{Values to replace.}

\item{RE_NAME}{Logical, rename the replaced variable.}

\item{parallel}{Logical, parallel computing. Default is TRUE.}

\item{x}{Name of variable to replace value.}
}
\description{
\code{replace_value} is for replacing values of some variables .
\code{replace_value_x} is for replacing values of a variable.
}
