% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_proc_tools.R
\name{require_packages}
\alias{require_packages}
\title{require_packages}
\usage{
require_packages(pkg)
}
\arguments{
\item{pkg}{A list or vector of names of required packages.}
}
\value{
packages installed and library.
}
\description{
\code{require_packages} is function for librarying required packages and  installing missing packages if needed.
}
\examples{
\dontrun{
require_packages(c("data.table","ggplot2"))
}
}
