% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_anaylsis.R
\name{rule_value_replace}
\alias{rule_value_replace}
\title{rule_value_replace}
\usage{
rule_value_replace(
  dat,
  rules_list,
  VALUE = 1:length(rules_list),
  x_name = "x_level"
)
}
\arguments{
\item{dat}{A data.frame.}

\item{rules_list}{Names of variables to replace value.}

\item{VALUE}{values to replace.}

\item{x_name}{name of the new variable.}
}
\description{
\code{rule_value_replace} is for generating new variables by rules.
}
