% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process_tools.R
\name{save_data}
\alias{save_data}
\title{Save data}
\usage{
save_data(
  ...,
  files = list(...),
  file_name = as.character(substitute(list(...))),
  dir_path = getwd(),
  note = FALSE,
  as_list = FALSE,
  row_names = FALSE,
  append = FALSE
)
}
\arguments{
\item{...}{datasets}

\item{files}{A dataset or a list of datasets.}

\item{file_name}{The file name of data.}

\item{dir_path}{A string. The dir path to save breaks_list.}

\item{note}{Logical. Outputs info.Default is TRUE.}

\item{as_list}{Logical. List format or data.frame format to save. Default is FALSE.}

\item{row_names}{Logical,retain rownames.}

\item{append}{Logical, append newdata to old.}
}
\description{
\code{save_data} is for saving a data.frame or a list fast.
}
\examples{
save_data(UCICreditCard,"UCICreditCard", tempdir())
}
