% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process_tools.R
\name{save_dt}
\alias{save_dt}
\title{Save data}
\usage{
save_dt(dt, file_name = "dat", dir_path = getwd(), note = FALSE,
  as_list = FALSE, row_names = FALSE, append = FALSE)
}
\arguments{
\item{dt}{A data frame or list.}

\item{file_name}{A string. The file name to save breaks_list.}

\item{dir_path}{A string. The dir path to save breaks_list.}

\item{note}{Logical. Outputs info.Default is TRUE.}

\item{as_list}{Logical. List format or data.frame format to save. Default is FALSE.}

\item{row_names}{Logical,retain rownames.}

\item{append}{Logical, append newdata to old.}
}
\description{
\code{save_dt} is for saving a data.frame or a list fast.
}
\examples{
save_dt(UCICreditCard,"UCICreditCard", tempdir())
}
