% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_score_card.R
\name{score_transfer}
\alias{score_transfer}
\title{Score Transfer}
\usage{
score_transfer(model, tbl_woe, a = 600, b = 50,
  file_name = "dat_score", dir_path = "./data/", save_data = TRUE)
}
\arguments{
\item{model}{A data frame with x and target.}

\item{tbl_woe}{a data.frame with woe variables.}

\item{a}{Base line of score.}

\item{b}{Numeric.Increased scores from doubling Odds.}

\item{file_name}{The name for periodically saved score file. Default is "dat_score".}

\item{dir_path}{The path for periodically saved score file.  Default is "./data"}

\item{save_data}{Logical, save results in locally specified folder. Default is TRUE}
}
\value{
A data.frame with variables which values transfered to score.
}
\description{
\code{score_transfer} is  for transfer woe to score.
}
\examples{
\dontrun{
score_transfer(model, tbl_woe, a = 600, b = 50, 
file_name = "dat_score", dir_path = "./data/", save_data = TRUE)
}
}
