% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_derivation_tools.R
\name{sim_str}
\alias{sim_str}
\title{sim_str}
\usage{
sim_str(a, b, sep = "_|[.]|[A-Z]")
}
\arguments{
\item{a}{A string}

\item{b}{A string}

\item{sep}{Seprater of strings. Default is "_|[.]|[A-Z]".}
}
\description{
This function is not intended to be used by end user.
}
