% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_binning.R
\name{split_bins}
\alias{split_bins}
\title{split_bins}
\usage{
split_bins(dat, x, breaks = NULL, bins_no = TRUE)
}
\arguments{
\item{dat}{A data.frame with independent variables.}

\item{x}{The name of an independent variable.}

\item{breaks}{Breaks for binning.}

\item{bins_no}{Number the generated bins. Default is TRUE.}
}
\value{
A data.frame with Bined x.
}
\description{
\code{split_bins} is  for binning using breaks.
}
\examples{
bins = split_bins(dat = UCICreditCard,
x = "PAY_AMT1", breaks = NULL, bins_no = TRUE)
}
