% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_parallel_computing.R
\name{start_parallel_computing}
\alias{start_parallel_computing}
\title{Parallel computing and export variables to global Env.}
\usage{
start_parallel_computing(parallel = TRUE)
}
\arguments{
\item{parallel}{A logical, default is TRUE.}
}
\value{
parallel works.
}
\description{
Parallel computing and export variables to global Env.
}
\examples{
\dontrun{
parallel = start_parallel_computing(parallel)
}
}
