% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process_tools.R
\name{stop_parallel_computing}
\alias{stop_parallel_computing}
\title{Stop parallel computing}
\usage{
stop_parallel_computing(cluster)
}
\arguments{
\item{cluster}{Parallel works.}
}
\value{
stop clusters.
}
\description{
This function  is not intended to be used by end user.
}
