% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process_tools.R
\name{str_match}
\alias{str_match}
\title{string match
#' \code{str_match} search for matches to argument pattern within each element of a character vector:}
\usage{
str_match(pattern, str_r)
}
\arguments{
\item{pattern}{character string containing a regular expression (or character string for fixed = TRUE) to be matched in the given character vector. Coerced by as.character to a character string if possible. If a character vector of length 2 or more is supplied, the first element is used with a warning. missing values are allowed except for regexpr and gregexpr.}

\item{str_r}{a character vector where matches are sought, or an object which can be coerced by as.character to a character vector. Long vectors are supported.}
}
\description{
string match
#' \code{str_match} search for matches to argument pattern within each element of a character vector:
}
\examples{
orignal_nam = c("12mdd","11mdd","10mdd")
str_match(str_r = orignal_nam,pattern= "\\\\d+")
}
