% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exploration.R
\name{sum_table}
\alias{sum_table}
\title{Summary table}
\usage{
sum_table(dat, ..., x_s = as.character(substitute(list(...))), x_list = NULL)
}
\arguments{
\item{dat}{A data.frame with x and target.}

\item{...}{x of dat}

\item{x_s}{A list of x.}

\item{x_list}{Names of dat.}
}
\value{
A list contains both categrory and numeric variable analysis.
}
\description{
#'The \code{sum_table} includes both univariate and bivariate analysis and ranges from univariate statistics and frequency distributions, to correlations, cross-tabulation and characteristic analysis.
}
\examples{
sum_table(UCICreditCard)
sum_table(UCICreditCard,LIMIT_BAL,AGE,EDUCATION,SEX)
}
