% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{time_series_proc}
\alias{time_series_proc}
\title{Process time series data}
\usage{
time_series_proc(dat, ID = NULL, group = NULL, time = NULL)
}
\arguments{
\item{dat}{A data.frame contained only predict variables.}

\item{ID}{The name of ID of observations or key variable of data. Default is NULL.}

\item{group}{The group of behavioral or status variables.}

\item{time}{The name of variable which is time when behavior was happened.}
}
\description{
This function is used for time series data processing.
}
\details{
The key to creating a good model is not the power of a specific modelling technique, but the breadth and depth of derived variables that represent a higher level of knowledge about the phenomena under examination.
}
\examples{
dat = data.frame(id = c(1,1,1,2,2,3,3,3,4,4,4,4,4,5,5,6,7,7,
                            8,8,8,9,9,9,10,10,11,11,11,11,11,11),
                     terms = c('a','b','c','a','c','d','d','a',
                               'b','c','a','c','d','a','c',
                                  'd','a','e','f','b','c','f','b',
                               'c','h','h','i','c','d','g','k','k'),
                     time = c(8,3,1,9,6,1,4,9,1,3,4,8,2,7,1,
                              3,4,1,8,7,2,5,7,8,8,2,1,5,7,2,7,3))

time_series_proc(dat = dat, ID = 'id', group = 'terms',time = 'time')
}
