% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{time_variable}
\alias{time_variable}
\title{time_variable}
\usage{
time_variable(dat, date_cols = NULL, enddate = NULL)
}
\arguments{
\item{dat}{A data.frame.}

\item{date_cols}{Time variables.}

\item{enddate}{End time.}
}
\description{
This function is not intended to be used by end user.
}
