% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_derivation.R
\name{time_varieble}
\alias{time_varieble}
\title{time_varieble}
\usage{
time_varieble(dat, date_cols = NULL, enddate = NULL)
}
\arguments{
\item{dat}{A data.frame.}

\item{date_cols}{Time variables.}

\item{enddate}{End time.}
}
\description{
This function is not intended to be used by end user.
}
