% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_derivation.R
\name{time_vars_process}
\alias{time_vars_process}
\title{Processing of Time or Date Variables}
\usage{
time_vars_process(df_tm = df_tm, x, enddate = "occur_time")
}
\arguments{
\item{df_tm}{A data.frame}

\item{x}{Time variable.}

\item{enddate}{End time.}
}
\description{
This function is not intended to be used by end user.
}
