% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/essential_algorithms.R
\name{tnr_value}
\alias{tnr_value}
\title{tnr_value}
\usage{
tnr_value(prob, target)
}
\arguments{
\item{prob}{A list of redict probability or score.}

\item{target}{Vector of target.}
}
\value{
True Positive Rate
}
\description{
\code{tnr_value} is for get true negtive rate for a prob or score.
}
