% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_training.R
\name{training_model}
\alias{training_model}
\title{Training model}
\usage{
training_model(model_name = "model", dat_train, dat_test = NULL,
  target, pos_flag = NULL, x_list = NULL, prop = 0.7,
  occur_time = NULL, obs_id = NULL, miss_values = NULL,
  ex_cols = NULL, preproc = TRUE, outlier_proc = TRUE,
  missing_proc = TRUE, default_miss = FALSE,
  feature_filter = list(filter = c("IV", "PSI", "COR", "XGB"), cv_folds =
  1, iv_cp = 0.02, psi_cp = 0.1, xgb_cp = 0, hopper = FALSE),
  algorithm = list("LR", "XGB"), LR.params = lr_params(),
  XGB.params = xgb_params(), GBM.params = gbm_params(),
  RF.params = rf_params(), breaks_list = NULL, parallel = FALSE,
  cores_num = NULL, save_pmml = TRUE, plot_show = FALSE, seed = 46,
  ...)
}
\arguments{
\item{model_name}{A string, name of the project. Default is NULL..}

\item{dat_train}{A data.frame with independent variables and target variable.}

\item{dat_test}{A data.frame of test data. Default is NULL.}

\item{target}{The name of target variable.}

\item{pos_flag}{The value of positive class of target variable, default: "1".}

\item{x_list}{Names of independent variables. Default is NULL.}

\item{prop}{Percentage of train-data after the partition. Default: 0.7.}

\item{occur_time}{The name of the variable that represents the time at which each observation takes place.Default is NULL.}

\item{obs_id}{The name of ID of observations or key variable of data. Default is NULL.}

\item{miss_values}{Other extreme value might be used to represent missing values, e.g: -9999, -9998. These miss_values will be encoded to -1 or "Unknown".}

\item{ex_cols}{Names of excluded variables. Regular expressions can also be used to match variable names. Default is NULL.}

\item{preproc}{Logical. Preprocess data. Default is TRUE}

\item{outlier_proc}{Logical. If TRUE,  Outliers processing using Kmeans and Local Outlier Factor. Default is TRUE}

\item{missing_proc}{Logical. If TRUE, missing value analysis and process missing value by knn imputation or central impulation or random imputation. Default is TRUE}

\item{default_miss}{Logical. If TRUE, assigning the missing values to -1 or "Unknown", otherwise, processing the missing values according to the results of missing analysis. See details at: \code{\link{process_nas}}}

\item{feature_filter}{Parameters for selecting important and stable features.See details at: \code{\link{feature_select_wrapper}}}

\item{algorithm}{Algorithms for training a model. list("LR", "XGB", "GBDT", "RF") are available.}

\item{LR.params}{Parameters of logistic regression & scorecard. See details at :  \code{\link{lr_params}}. 
\itemize{
  \item \code{tree_control} the list of parameters to control cutting initial breaks by decision tree. See details at: \code{\link{get_tree_breaks}}
  \item \code{bins_control} the list of parameters to control merging initial breaks. See details at: \code{\link{select_best_breaks}},\code{\link{select_best_class}}
  \item \code{best_lambda} Metheds of best lanmbda stardards using to filter variables by LASSO.There are four methods: ("lambda.min", "lambda.1se", "lambda.05se" , "lambda.sim_sign") . Default is  "lambda.sim_sign". See details at: \code{\link{get_best_lambda}}
  \item \code{obsweight} An optional vector of  'prior weights' to be used in the fitting process. Should be NULL or a numeric vector. If you oversample or cluster diffrent datasets to training the LR model, you need to set this parameter to ensure that the probability of logistic regression output is the same as that before oversampling or segmentation. e.g.:There are 10,000 good obs and 500 bad obs before oversampling or under-sampling, 5,000 good obs and 3,000 bad obs after oversampling. Then this parameter should be set to c(10000/5000, 500/3000). Default is NULL..
  \item \code{forced_in}Names of forced input variables. Default is NULL.
  \item \code{sp_values} Vaules will be in separate bins.e.g. list(-1, "Unknown")  means that -1 & Unknown as special values.Default is NULL.
  \item \code{step_wise} Logical, stepwise method. Default is TRUE.
  \item \code{score_card} Logical, transfer woe to a standard scorecard. If TRUE, Output scorecard, and score prediction, otherwise output probability. Default is TRUE.
  \item \code{cor_p} The maximum threshold of correlation.0 <= cor_p <=1; 0.5 to 0.8 usually work.  Default: 0.7.
  \item \code{iv_i} The minimum threshold of IV. 0 < iv_i ; 0.01 to 0.1 usually work. Default: 0.01
  \item \code{psi_i} The maximum threshold of PSI.  0 <= psi_i <=1; 0.05 to 0.2 usually work. Default: 0.1
}}

\item{XGB.params}{Parameters of xgboost. See details at :  \code{\link{xgb_params}}.}

\item{GBM.params}{Parameters of GBM. See details at :  \code{\link{gbm_params}}.}

\item{RF.params}{Parameters of Random Forest. See details at :  \code{\link{rf_params}}.}

\item{breaks_list}{A table containing a list of splitting points for each independent variable. Default is NULL.}

\item{parallel}{Default is FALSE}

\item{cores_num}{The number of CPU cores to use.}

\item{save_pmml}{Logical, save model in PMML format. Default is TRUE.}

\item{plot_show}{Logical, show model performance in current graphic device. Default is FALSE.}

\item{seed}{Random number seed. Default is 46.}

\item{...}{Other parameters.}
}
\value{
A list containing Model Objects.
}
\description{
\code{training_model} Model builder
}
\examples{
\dontrun{
B_model = training_model(dat_train = lendingclub,
model_name = "lendingclub",target = "loan_status",
occur_time = "issue_d", obs_id = "id", prop = 0.7, 
algorithm = list("LR", "XGB"),seed = 46)
}
}
\seealso{
\code{\link{train_test_split}},\code{\link{cleaning_data}}, \code{\link{feature_select_wrapper}},   \code{\link{lr_params}}, \code{\link{xgb_params}}, \code{\link{gbm_params}}, \code{\link{rf_params}},\code{\link{fast_high_cor_filter}},\code{\link{get_breaks_all}},\code{\link{lasso_filter}}, \code{\link{woe_trans_all}}, \code{\link{get_logistic_coef}}, \code{\link{score_transfer}},\code{\link{get_score_card}}, \code{\link{model_key_index}},\code{\link{ks_psi_plot}},\code{\link{get_plots}},\code{\link{ks_table_plot}}
}
