% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_derivation.R
\name{variable_process}
\alias{variable_process}
\title{variable_process}
\usage{
variable_process(add)
}
\arguments{
\item{add}{A data.frame}
}
\description{
This function is not intended to be used by end user.
}
