% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vintage_analysis.R
\name{vintage_function}
\alias{vintage_function}
\title{vintage_function
\code{vintage_function} is for vintage analysis.}
\usage{
vintage_function(dat, obs_id = NULL, occur_time = NULL, MOB = NULL,
  period = "monthly", status = NULL, amount = NULL, by_out = "cnt",
  start_date = NULL, end_date = NULL, dead_status = 30)
}
\arguments{
\item{dat}{A data.frame contained id, occur_time, mob, status ...}

\item{obs_id}{The name of ID of observations or key variable of data. Default is NULL.}

\item{occur_time}{The name of the variable that represents the time at which each observation takes place.}

\item{MOB}{Mobility of book}

\item{period}{Period of event to analysiss. Default is "monthly"}

\item{status}{Status of observations}

\item{amount}{The name of variable representing amount. Default is NULL.}

\item{by_out}{Output: amount (amt) or count (cnt)}

\item{start_date}{The earliest occurrence time of observations.}

\item{end_date}{The latest occurrence time of observations.}

\item{dead_status}{Status of dead observations.}
}
\description{
This function is not intended to be used by end user.
}
