% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_training.R
\name{xgb_data}
\alias{xgb_data}
\title{XGboost data}
\usage{
xgb_data(
  dat_train,
  target,
  dat_test = NULL,
  x_list = NULL,
  prop = 0.7,
  occur_time = NULL
)
}
\arguments{
\item{dat_train}{data.frame of train data. Default is NULL.}

\item{target}{name of target variable.}

\item{dat_test}{data.frame of test data. Default is NULL.}

\item{x_list}{names of independent variables of raw data. Default is NULL.}

\item{prop}{Percentage of train-data after the partition. Default: 0.7.}

\item{occur_time}{The name of the variable that represents the time at which each observation takes place.Default is NULL.}
}
\description{
\code{xgb_data} is for prepare data using in \code{\link{training_model}}.
}
