% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process_tools.R
\name{as_percent}
\alias{as_percent}
\title{Percent Format}
\usage{
as_percent(x, digits = 2)
}
\arguments{
\item{x}{A numeric vector or  list.}

\item{digits}{Number of digits.Default: 2.}
}
\value{
x with percent format.
}
\description{
\code{as_percent} is  a small function for making percent format..
}
\examples{
as_percent(0.2363, digits = 2)
as_percent(1)
}
