% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{char_to_num}
\alias{char_to_num}
\title{character to number}
\usage{
char_to_num(
  dat,
  char_list = NULL,
  m = 0,
  p = 0.5,
  note = FALSE,
  ex_cols = NULL
)
}
\arguments{
\item{dat}{A data frame}

\item{char_list}{The list of charecteristic variables that need to merge categories, Default is NULL. In case of NULL, merge categories for all variables of string type.}

\item{m}{The minimum number of categories.}

\item{p}{The max percent of categories.}

\item{note}{Logical, outputs info. Default is TRUE.}

\item{ex_cols}{A list of excluded variables. Regular expressions can also be used to match variable names. Default is NULL.}
}
\value{
A data.frame
}
\description{
\code{char_to_num} is  for transfering character variables which are actually numerical numbers containing strings  to numeric.
}
\examples{
dat_sub = lendingclub[c('dti_joint',	'emp_length')]
str(dat_sub)
#variables that are converted to numbers containing strings
dat_sub = char_to_num(dat_sub)
str(dat_sub)
}
