% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_proc_tools.R
\name{char_to_num}
\alias{char_to_num}
\title{character to number}
\usage{
char_to_num(dat, note = FALSE,
  ex_cols = "date$|id$|time$|DATA$|ID$|TIME$")
}
\arguments{
\item{dat}{A data frame}

\item{note}{Logical, outputs info. Default is TRUE.}

\item{ex_cols}{A list of excluded variables. Regular expressions can also be used to match variable names. Default is NULL.}
}
\value{
A data.frame
}
\description{
\code{char_to_num} is  for transfering character variables which are actually numerical to numeric.
}
\examples{
\dontrun{
char_to_num(lendingclub,ex_cols = "id$|_d$")
}
}
