% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/essential_algorithms.R
\name{cos_sim}
\alias{cos_sim}
\title{cos_sim}
\usage{
cos_sim(x, y, cos_margin = 1)
}
\arguments{
\item{x}{A list of numbers}

\item{y}{A list of numbers}

\item{cos_margin}{Margin of matrix, 1 for rows and 2 for cols, Default is 1.}
}
\value{
A number of cosin similarity
}
\description{
This function is not intended to be used by end user.
}
