% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_customer_segmentation.R
\name{cos_sim}
\alias{cos_sim}
\title{cos_sim}
\usage{
cos_sim(x, y, margin = 1)
}
\arguments{
\item{x}{A list}

\item{y}{A list}

\item{margin}{rows or cols}
}
\description{
This function is not intended to be used by end user.
}
