% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_customer_segmentation.R
\name{fuzzy_cluster_means}
\alias{fuzzy_cluster_means}
\alias{fuzzy_cluster}
\title{Fuzzy Cluster means.}
\usage{
fuzzy_cluster_means(dat, kc = 2, sf = 2, nstart = 1,
  max_iter = 100, epsm = 1e-06)

fuzzy_cluster(dat, kc = 2, init_centers, sf = 3, max_iter = 100,
  epsm = 1e-06)
}
\arguments{
\item{dat}{A data.frame contained only predict variables.}

\item{kc}{The number of cluster center (default is 2),}

\item{sf}{Default is 2.}

\item{nstart}{The number of random groups (default is 1),}

\item{max_iter}{Max iteration number(default is 100) .}

\item{epsm}{Default is 1e-06.}

\item{init_centers}{Initial centers of obs.}
}
\description{
This function is used for Fuzzy Clustering.
}
