% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_filter_iv_xgb.R
\name{gbm_filter}
\alias{gbm_filter}
\title{Select Features using GBM}
\usage{
gbm_filter(dat, target = "flag", x_list = NULL, ex_cols = NULL,
  pos_flag = NULL, gbdt_params = list(n.trees = 1000, interaction.depth
  = 6, shrinkage = 0.01, bag.fraction = 0.5, train.fraction = 0.7,
  n.minobsinnode = 30, cv.folds = 5, best_iter = TRUE, method = "cv"),
  cores_num = 2, vars_name = TRUE, note = TRUE, save_data = FALSE,
  file_name = "Featrue_importance_GBDT", dir_path = "./variable",
  seed = 46, ...)
}
\arguments{
\item{dat}{A data.frame with independent variables and target variable.}

\item{target}{The name of target variable.}

\item{x_list}{Names of independent variables.}

\item{ex_cols}{A list of excluded variables. Regular expressions can also be used to match variable names. Default is NULL.}

\item{pos_flag}{The value of positive class of target variable, default: "1".}

\item{gbdt_params}{Parameters of GBM.The complete list of parameters is available at: \code{\link{gbm}}.}

\item{cores_num}{The number of CPU cores to use.}

\item{vars_name}{Logical, output a list of filtered variables or table with detailed IV and PSI value of each variable. Default is TRUE.}

\item{note}{Logical, outputs info. Default is TRUE.}

\item{save_data}{Logical, save results results in locally specified folder. Default is TRUE}

\item{file_name}{The name for periodically saved results files.  Default is "Featrue_importance_GBDT".}

\item{dir_path}{The path for periodically saved results files.  Default is "./variable".}

\item{seed}{Random number seed. Default is 46.}

\item{...}{Other parameters to pass to gbdt_params.}
}
\value{
Selected variables.
}
\description{
\code{gbm_filter}  is for selecting important features using GBM.
}
\examples{
\dontrun{
gbm_filter(dat = UCICreditCard, target = "default.payment.next.month", 
occur_time = "apply_date", 
ex_cols = "ID$|date$|default.payment.next.month$", 
gbdt_params = list(n.trees = 100, interaction.depth = 6, shrinkage = 0.01, 
bag.fraction = 0.5, train.fraction = 0.7, n.minobsinnode = 30, 
cv.folds = 5, best_iter = TRUE, method = "cv"),
vars_name = FALSE)
}
}
\seealso{
\code{\link{psi_iv_filter}}, \code{\link{xgb_filter}}, \code{\link{feature_select_wrapper}}
}
