% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{ks_table}
\alias{ks_table}
\alias{ks_table_plot}
\alias{ks_psi_plot}
\title{ks_table & plot}
\usage{
ks_table(train_pred, test_pred, target = NULL, score = "score",
  g = 10, breaks = NULL, pos_flag = NULL)

ks_table_plot(train_pred, test_pred, target = "target",
  score = "score", g = 10, plot_show = TRUE, g_width = 13,
  dir_path = "./pic", gtitle = NULL)

ks_psi_plot(train_pred, test_pred, target = "target", score = "score",
  gtitle = NULL, plot_show = TRUE, g_width = 12, breaks = NULL,
  g = 10, dir_path = "./pic")
}
\arguments{
\item{train_pred}{A data frame of training with predicted prob or score.}

\item{test_pred}{A data frame of validation with predict prob or score.}

\item{target}{The name of target variable.}

\item{score}{The name of prob or score variable.}

\item{g}{Number of breaks for prob or score.}

\item{breaks}{Splitting points of prob or score.}

\item{pos_flag}{The value of positive class of target variable, default: "1".}

\item{plot_show}{Logical, show model performance in current graphic device. Default is FALSE.}

\item{g_width}{Width of graphs.}

\item{dir_path}{The path for periodically saved graphic files. Default is "./pic"}

\item{gtitle}{The title of the graph & The name for periodically saved graphic file. Default is "_ks_psi_table".}
}
\description{
\code{ks_table} is for generating a model performance table.
\code{ks_table_plot} is for ploting the table generated by \code{ks_table}
\code{ks_psi_plot} is for K-S & PSI distrbution ploting.
}
\examples{
\dontrun{
ks_table(train_pred = dat_train, test_pred =dat_test, 
target = "flag", score = "score", g = 10)
}
}
