% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{ks_value}
\alias{ks_value}
\title{ks_value}
\usage{
ks_value(dat_pred = NULL, target = NULL, score = NULL, g = 20,
  breaks = NULL)
}
\arguments{
\item{dat_pred}{A data frame with predict prob or score.}

\item{target}{The name of target variable.}

\item{score}{The name of prob or score variable.}

\item{g}{Number of breaks for prob or score.}

\item{breaks}{Splitting points of prob or score.}
}
\description{
\code{ks_value} is for get plots for a  variable.
}
\examples{
\dontrun{
ks_value(dat_pred = dat_train,
target = "flag", score ="score", g = 20)
}
}
