% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualization.R
\name{love_color}
\alias{love_color}
\title{love_color}
\usage{
love_color(color = NULL, type = NULL, all = FALSE)
}
\arguments{
\item{color}{The name of colors.}

\item{type}{The type of colors, "deep".}

\item{all}{all of colors.}
}
\description{
\code{love_color} is for get plots for a  variable.
}
\examples{
love_color(color="dark_cyan")
}
