% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_proc_tools.R
\name{min_max_norm}
\alias{min_max_norm}
\title{min_max_norm}
\usage{
min_max_norm(x)
}
\arguments{
\item{x}{A data.frame or Matrix.}
}
\value{
A list
}
\description{
\code{min_max_norm} Functions for vector operation.
}
\examples{
\dontrun{
dat = one_hot_encoding(dat = UCICreditCard, ex_col = "ID$|_date$",merge_cat = TRUE, na_act = TRUE)
dat = low_variance_filter(dat, lvp = 0.9)
dat_s = apply(dat, 2, min_max_norm)
}
}
