% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualization.R
\name{plot_box}
\alias{plot_box}
\title{Plot Box}
\usage{
plot_box(
  dat,
  y,
  x = NULL,
  x_breaks = NULL,
  g = 5,
  fill_colors = c(love_color(type = "lightnihon_8x1"), love_color(type =
    "darknihon_8x1"))
)
}
\arguments{
\item{dat}{A data.frame with independent variables and target variable.}

\item{y}{The name of target variable.}

\item{x}{The name of an independent variable.}

\item{x_breaks}{Breaks points of x.}

\item{g}{Number of initial breakpoints for equal frequency binning.}

\item{fill_colors}{colors of x.}
}
\description{
You can use the \code{plot_box} to produce boxplot.
}
\examples{
plot_box(lendingclub, x = "grade", y = "installment", g = 7)
}
