% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_training.R
\name{pred_score}
\alias{pred_score}
\title{pred_score}
\usage{
pred_score(
  model,
  dat,
  x_list = NULL,
  bins_table = NULL,
  obs_id = NULL,
  miss_values = list(-1, "-1", "NULL", "-1", "-9999", "-9996", "-9997", "-9995",
    "-9998", -9999, -9998, -9997, -9996, -9995),
  woe_name = FALSE
)
}
\arguments{
\item{model}{Logistic Regression Model generated by \code{\link{training_model}}.}

\item{dat}{Dataframe of new data.}

\item{x_list}{Into the model variables.}

\item{bins_table}{a data.frame generated by \code{\link{get_bins_table}}}

\item{obs_id}{The name of ID of observations or key variable of data. Default is NULL.}

\item{miss_values}{Special values.}

\item{woe_name}{Logical. Whether woe variable's name contains 'woe'.Default is FALSE.}
}
\value{
new scores.
}
\description{
\code{pred_score} is for using logistic regression model model to predict new data.
}
\seealso{
\code{\link{training_model}}, \code{\link{lr_params}}, \code{\link{xgb_params}}, \code{\link{rf_params}}
}
