% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_proc_tools.R
\name{rowAny}
\alias{rowAny}
\alias{rowAllnas}
\alias{colAllnas}
\alias{colAllzeros}
\alias{rowAll}
\alias{rowCVs}
\alias{rowSds}
\alias{colSds}
\alias{rowMaxs}
\alias{rowMins}
\alias{rowMaxMins}
\title{Functions for vector operation.}
\usage{
rowAny(x)

rowAllnas(x)

colAllnas(x)

colAllzeros(x)

rowAll(x)

rowCVs(x, na.rm = FALSE)

rowSds(x, na.rm = FALSE)

colSds(x, na.rm = FALSE)

rowCVs(x, na.rm = FALSE)

rowSds(x, na.rm = FALSE)

rowMaxs(x, na.rm = FALSE)

rowMins(x, na.rm = FALSE)

rowMaxMins(x, na.rm = FALSE)
}
\arguments{
\item{x}{A data.frame or Matrix.}

\item{na.rm}{Logical, remove NAs.}
}
\value{
A list
}
\description{
Functions for vector operation.
}
\examples{
\dontrun{
 rowAny(UCICreditCard[8:10])
 rowAllnas(UCICreditCard)
}
}
