% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_binning.R
\name{split_bins}
\alias{split_bins}
\title{split_bins}
\usage{
split_bins(
  dat,
  x,
  breaks = NULL,
  bins_no = TRUE,
  as_factor = FALSE,
  labels = NULL,
  use_NA = TRUE,
  char_free = FALSE
)
}
\arguments{
\item{dat}{A data.frame with independent variables.}

\item{x}{The name of an independent variable.}

\item{breaks}{Breaks for binning.}

\item{bins_no}{Number the generated bins. Default is TRUE.}

\item{as_factor}{Whether to convert to factor type.}

\item{labels}{Labels of bins.}

\item{use_NA}{Whether to process NAs.}

\item{char_free}{Logical, if TRUE, characters are not splitted.}
}
\value{
A data.frame with Bined x.
}
\description{
\code{split_bins} is  for binning using breaks.
}
\examples{
bins = split_bins(dat = UCICreditCard,
x = "PAY_AMT1", breaks = NULL, bins_no = TRUE)
}
