% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process_tools.R
\name{start_parallel_computing}
\alias{start_parallel_computing}
\title{Parallel computing and export variables to global Env.}
\usage{
start_parallel_computing(parallel = TRUE)
}
\arguments{
\item{parallel}{A logical, default is TRUE.}
}
\value{
parallel works.
}
\description{
This function  is not intended to be used by end user.
}
