% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_parallel_computing.R
\name{stop_parallel_computing}
\alias{stop_parallel_computing}
\title{Stop parallel computing}
\usage{
stop_parallel_computing(cluster)
}
\arguments{
\item{cluster}{Parallel works.}
}
\value{
stop clusters.
}
\description{
Stop parallel computing
}
\examples{
\dontrun{
parallel = start_parallel_computing(parallel)
stop_parallel_computing(attr(parallel, "cluster"))
}
}
