% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{time_series_proc}
\alias{time_series_proc}
\title{Process time series data}
\usage{
time_series_proc(dat, ID = NULL, group = NULL, time = NULL)
}
\arguments{
\item{dat}{A data.frame contained only predict variables.}

\item{ID}{The name of ID of observations or key variable of data. Default is NULL.}

\item{group}{The group of behavioral or status variables.}

\item{time}{The name of variable which is time when behavior was happened.}
}
\description{
This function is used for time series data processing.
}
\details{
The key to creating a good model is not the power of a specific modelling technique, but the breadth and depth of derived variables that represent a higher level of knowledge about the phenomena under examination.
}
