% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_woe_transfer.R
\name{woe_trans_all}
\alias{woe_trans_all}
\alias{woe_trans}
\title{Converting data to WOE}
\usage{
woe_trans_all(dat, x_list = NULL, ex_cols = NULL, bins_table = NULL,
  target = NULL, breaks_list = NULL, note = TRUE, save_data = TRUE,
  parallel = FALSE, woe_name = TRUE, file_name = "dat_woe",
  dir_path = "./data", ...)

woe_trans(dat, x, bins_table = NULL, target = NULL,
  breaks_list = NULL, woe_name = TRUE)
}
\arguments{
\item{dat}{A data.frame with independent variables.}

\item{x_list}{A list of x variables.}

\item{ex_cols}{Names of excluded variables. Regular expressions can also be used to match variable names. Default is NULL.}

\item{bins_table}{A table contians woe of each bin of variables, it is generated by code{\link{get_bins_table_all}},code{\link{get_bins_table}}}

\item{target}{The name of target variable. Default is NULL.}

\item{breaks_list}{A list contains breaks of variables. it is generated by code{\link{get_breaks_all}},code{\link{get_breaks}}}

\item{note}{Logical, outputs info. Default is TRUE.}

\item{save_data}{Logical, save results in locally specified folder. Default is TRUE}

\item{parallel}{Logical, parallel computing. Default is FALSE.}

\item{woe_name}{Logical. Add "_woe" at the end of the variable name.}

\item{file_name}{The name for periodically saved woe file. Default is "dat_woe".}

\item{dir_path}{The path for periodically saved woe file Default is "./data"}

\item{...}{Additional parameters.}

\item{x}{The name of an independent variable.}
}
\value{
A list of breaks for each variables.
}
\description{
\code{woe_trans} is for converting dat to woe.
The \code{woe_trans_all} function is a simpler wrapper for \code{woe_trans}.
}
\examples{
\dontrun{
# load germancredit dat
dat(UCI_Credit_Card)
woe_trans_all(dat = UCICreditCard,
x_list = names(UCICreditCard)[3:10],
target = "default.payment.next.month",
occur_time = "apply_date",  ex_cols = "ID|apply_date")
woe_transUCICreditCard, x = "PAY_3", 
target = "default.payment.next.month")
}

}
\seealso{
\code{\link{get_tree_breaks}}, \code{\link{cut_equal}}, \code{\link{select_best_class}}, \code{\link{select_best_breaks}}
}
